Add Listener#unregister() method to remove it
I'm not sure how EventEmitter#removeListener knows how to check the
passed callback against the callback to remove in its internal array, so
to be safe, I saved the bound listener callback as a variable and then
pass that into Client#on/Client#once and EventEmitter#removeListener.
The bot still runs, but I'm not sure if unregister() works since I have
not tested it.