Allow listener callbacks to return boolean success
The naive flow of always refreshing cooldowns following void callbacks
(provided they don't error) makes it so that if a callback runs some
filtering of its own and decides not to do what should be counted
towards the cooldown, the cooldown is refreshed anyway. For example, in
LukeService#processDadJoke, it's not guaranteed that it recognizes the
message as a dad joke trigger and responds with a dad joke. Listener
callbacks can not return a boolean representing if they "succeeded" in
doing what they wanted to do (separate from failing due to error). For
backwards compatibility (and brevity), existing void callbacks are
treated as always succeeding.