Port Coffee and Deez modules to new architecture
These included message listeners with and without cooldown, so this
involved refactoring CooldownManager. Cooldown is now treated as
middleware (thanks to the addition of the afterExecute feature giving
CooldownManager a chance to refresh cooldowns AFTER successful callback
execution). However, to support dynamic cooldown setting, the
CooldownManager object itself needs to be included on the spec such that
it can be retrieved from the client object at runtime.