Use new MockMessage, MockInteraction every test
This is needed because I couldn't get automatic resetting to work due
to different MockMessage#listener sharing the same ListenerSpec and thus
the same Cooldown (even after trying to use lodash.cloneDeep in various
places).

The solution for now was just to clone the ListenerSpec once on
MockMessage construction and give the burden of initializing different
MockMessages to the testers -- supposedly though, this shouldn't be too
much of an issue since tests are supposed to be isolated anyway, and
using a single MockMessage for all tests like before makes it *appear*
otherwise since the "automatic resetting" was transparent to the tester.
I made similar changes to MockInteraction just to be consistent.

However, this still seems like a code smell as I note in a code TODO:

TODO: A "spec" shouldn't have state saved on it but it does at the
moment in the form of CooldownManager. Thus, we have to clone the spec
for each instance of MockMessage to make sure tests using a distinct
MockMessage don't share cooldown state.

NOTE: Upon amending this commit, I realize this change doesn't even work
for some reason (listener still sharing cooldown) even though I swear I
saw it work earlier...? bruh