Fix bug of duplicate interaction replies
Previously, I tried to be clever by having CommandRunner#runExecute
check if the interaction was left not replied to and if so, reply with
a generic response. However, this failed to account for possible cleanup
callbacks that try to reply to the interaction, so when they try to
reply, it causes a Discord API error. This should be fixed by moving the
"fallback reply" code to the end of the top level CommandRunner#run
method.