Add per-channel option to /set-listener-cooldown
A problem has been identified in "preserving bypassers". When switching
between user and channel, since we treat IDs as the same, we get that a
user ID becomes interpreted as a channel ID, resulting in an invalid
bypasser when dumping. This doesn't affect underlying functionality,
just what the user sees. For example, the user ID -> channel ID cast
makes it such that we see #Unknown when inspecting with /cooldowns