Add custom footer for immunity/bypass embeds
Note that the tests converted all mockReturnValueOnce to
mockReturnValue. There was a bug where, because my new implementation
called timeoutService.isImmune() a separate time, it consumed the
expected return value before the test got to the part isImmune was
supposed to be mocked for. Since we reset mocks before every test
anyway, it would probably be a better idea to prefer the non-once
version going forward to prevent similar INFURIATING bugs.