Add missing await in /set-meow-chance
EMERGENCY FIX.

It was bringing down the entire bot. The missing await meant the
interaction was not guaranteed to be replied to before the command
runner queries the replied status. The command runner thus sometimes
sees that the interaction had not been replied to and uses the generic
response, only for the original reply to take effect and cause an
unknown interaction error. I could've sworn I fixed this bug a long time
ago, but I guess it was on a branch that never ended up getting merged.