Fix command runner tests
SO APPARENTLY, in ADDITION to clearMocks and resetMocks, there's also a
restoreMocks, which is for spies -- I think??. Now I have to go back and
figure out if I need set restoreMocks: true TOO after I can FINALLY fix
the resetMocks: true shit. JFC. jest.spyOn() on
CommandRunner#handleCommandError in the "checks" describe() was screwing
over the "checks > error handling" describe().

What was maddening was that when the "error handling" block was run
ALONE, everything passed. I had to figure out WHICH of the other
describes() was somehow leaking effects.