Remove test(Broadcast|Ephemeral)OptionSupport
They straight up don't work as expected. I have to go in and manually
add properties on the mock within each it() because for SOME reason the
invariants aren't being respected. I think this has something to do with
how within each test suite, we would regenerate the MockInteraction
or MockMessage with beforeEach(), but within the test...OptionSupport(),
we reuse the same mock object, which causes WEIRD behavior.