Copy assets into emitted code
EMERGENCY FIX. I haven't run the compiled version of the code in a long
time. This has been broken since I ported the Wordle feature (#97). tsc
doesn't emit static (non-TS) files, so the asset files weren't being
copied to dist/. This resulted in an error when starting up because the
Wordle module tries to load the Wordle dictionary text files.

This was actually a problem from way before, when assets/ was introduced
in the first place (#21), but I never noticed because of (1) rarely
running the compiled version of the code and (2) the listener trigger
being rarely used anyway.