Add TextChannelHandler
This was just to have a class for the last function remaining to be
refactored, `getDMChannel`.

Creating this class at all feels kind of forced with how we set up
HandlerProxy. Namely, it meant extending formatContext() to handle a
TextChannel overload, which may or may not make sense.

Furthermore, this is indicative of a potential architectural flaw.
formatContext() and HandlerProxy seem very coupled despite being in
completely separate modules. Since we make these handler classes to
abstract the handling of these discord.js objects in the first place,
maybe we can deprecate the original formatContext() later and move it to
be an implementation detail and responsibility of HandlerProxy.