Fix ^ shorthand to use ~, like in Git notation
Before, the notation treated ^N as equivalent to N ^'s. This notation is
just shorthand to use within the context of our own bot, so it doesn't
really matter in the sense that as long as we're consistent with what we
defined things to be, there's no correctness issues.

However, since this shorthand is inspired by and intended to mirror Git
revision notation, it would be better to be consistent with Git's
semantics. Namely, ^N in Git actually means the Nth parent and not
Nth-degree ancestor. It is ~N that is the shorthand for the latter,
being equivalent to N ^'s.