feat: add feature to have bot copy dev messages
The /set-concerted-react command has been repurposed to /set-concerted,
which handles both the setting for copying dev messages and the setting
for reacting to dev messages. It was also extended with another option,
and both options are optional --- if not provided, the bot will default
to the current setting as stored on the DevControlService.